% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api_helpers.R
\name{get_api_details}
\alias{get_api_details}
\title{Returns details for api config}
\usage{
get_api_details()
}
\value{
list with url and key
}
\description{
Returns details for api config
}
\examples{
get_api_details()
}

