% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_data.R
\name{get_heroes}
\alias{get_heroes}
\title{Retrieve current heroes details}
\usage{
get_heroes(api = get_api_details())
}
\arguments{
\item{api}{list from get_api_details}
}
\value{
api results in a list
}
\description{
Retrieve current heroes details
}
\examples{
get_heroes()
}

