% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_data.R
\name{get_item_details}
\alias{get_item_details}
\title{Retrieve details about a item}
\usage{
get_item_details(item_id, api = get_api_details())
}
\arguments{
\item{item_id}{a single item id}

\item{api}{list from get_api_details}
}
\value{
api results in a list
}
\description{
Retrieve details about a item
}
\examples{
get_item_details(item_id)
}

