% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/league_data.R
\name{get_league_listings}
\alias{get_league_listings}
\title{Retrieve league details}
\usage{
get_league_listings(api = get_api_details())
}
\arguments{
\item{api}{list from get_api_details}
}
\value{
api results in a list
}
\description{
Retrieve league details
}
\examples{
get_league_listings()
}

