% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/league_data.R
\name{get_live_league_games}
\alias{get_live_league_games}
\title{Retireve current live league games}
\usage{
get_live_league_games(api = get_api_details())
}
\arguments{
\item{api}{list from get_api_details}
}
\value{
api results in a list
}
\description{
Retireve current live league games
}
\examples{
get_live_league_games()
}

