% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/match_data.R
\name{get_match}
\alias{get_match}
\title{Retrieve basic match history}
\usage{
get_match(hero_id = NULL, game_mode = NULL, skill = NULL,
  min_players = NULL, account_id = NULL, league_id = NULL,
  start_at_match_id = NULL, matches_requested = NULL,
  tournament_games_only = NULL, api = get_api_details())
}
\arguments{
\item{hero_id}{Search for matches with a specific hero being played}

\item{game_mode}{Search for matches of a given mode}

\item{skill}{0 for any, 1 for normal, 2 for high, 3 for very high skill}

\item{min_players}{the minimum number of players required in the match}

\item{account_id}{Search for all matches for the given user}

\item{league_id}{matches for a particular league}

\item{start_at_match_id}{Start the search at the indicated match id, descending}

\item{matches_requested}{Maximum is 25 matches (default is 25)}

\item{tournament_games_only}{set to only show tournament games}

\item{api}{list from get_api_details}
}
\value{
api results in a list
}
\description{
Retrieve basic match history
}
\examples{
get_match()
}

